Combine QuadProcessor.cs and QuadProcessorUtility.cs
