--- src/QuadProcessor.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 27f008cf367d4df898b32024a745138c
-timeCreated: 1740445967
